fix(rides): crash on search/list — pg returns NUMERIC as string
ROUND(AVG()::numeric, 1) returns NUMERIC which the pg driver
deserializes as a JS string. Calling .toFixed() on it threw
\"toFixed is not a function\" and ErrorBoundary showed \"Algo salió mal\"
whenever the user opened search or any page rendering RideCard.

Backend (rides.js GET /):
- ROUND(...)::float → returns JS number
- COUNT(*)::int    → returns JS number

Frontend defense in depth — coerce to number before toFixed:
- RideCard:      Number(ride.driver_rating ?? 0).toFixed(1)
- TripInProgress: Number(rating ?? 0).toFixed(1)
- RateTrip:      Number(otherParty.rating ?? 0).toFixed(1)

Effect for users: trips were being saved correctly, but the search
results page was crashing on render, making it look like nothing
got saved. Both issues resolved by this single fix.

+++ backend/src/routes/rides.js
@@ -13,10 +13,10 @@ router.get('/', async (req, res) => {
            u.name as driver_name, u.phone as driver_phone,
            u.car_brand, u.car_color, u.car_plate,
            COALESCE(
-             (SELECT ROUND(AVG(rt.rating)::numeric, 1)
+             (SELECT ROUND(AVG(rt.rating)::numeric, 1)::float
               FROM ratings rt WHERE rt.ratee_id = u.id), 0
            ) as driver_rating,
-           (SELECT COUNT(*) FROM ratings rt WHERE rt.ratee_id = u.id) as driver_rating_count
+           (SELECT COUNT(*)::int FROM ratings rt WHERE rt.ratee_id = u.id) as driver_rating_count
     FROM rides r
     JOIN users u ON r.driver_id = u.id
     WHERE r.status = 'active' AND r.seats_available > 0
